requirements.txt
# Traot - Requirements
# Professional Trading Signal System v2.0

# =============================================================================
# CORE (pinned for stability)
# =============================================================================
numpy>=1.24.0,<2.0.0
pandas>=2.0.0,<3.0.0
pyyaml>=6.0,<7.0

# =============================================================================
# DATA COLLECTION
# =============================================================================
ccxt>=4.0.0,<5.0.0              # Exchange API (Binance, Coinbase, etc.)
requests>=2.31.0,<3.0.0         # HTTP requests (Twelve Data API, CoinGecko fallback)

# =============================================================================
# TECHNICAL ANALYSIS
# =============================================================================
scikit-learn>=1.3.0,<2.0.0      # Preprocessing, scaling

# =============================================================================
# MATHEMATICAL ENGINE (Advanced Algorithms)
# =============================================================================
scipy>=1.11.0,<2.0.0            # Scientific computing (Kalman, FFT, etc.)
PyWavelets>=1.4.0,<2.0.0        # Wavelet transform analysis

# =============================================================================
# MACHINE LEARNING - Unbreakable Ensemble System
# =============================================================================
torch>=2.0.0,<3.0.0             # PyTorch for TCN-LSTM-Attention model
xgboost>=2.0.0,<3.0.0           # XGBoost gradient boosting
lightgbm>=4.0.0,<5.0.0          # LightGBM gradient boosting
catboost>=1.2.0,<2.0.0          # CatBoost gradient boosting
ngboost>=0.4.0,<1.0.0           # NGBoost probabilistic predictions
joblib>=1.3.0,<2.0.0            # Model serialization

# =============================================================================
# NOTIFICATIONS
# =============================================================================
plyer>=2.1.0                    # Cross-platform desktop notifications
python-telegram-bot>=20.0,<22.0 # Telegram alerts (optional)

# =============================================================================
# DASHBOARD
# =============================================================================
streamlit>=1.28.0,<2.0.0        # Web dashboard
plotly>=5.17.0,<6.0.0           # Interactive charts
streamlit-autorefresh>=1.0.0    # Non-blocking auto-refresh for live updates
websocket-client>=1.6.0,<2.0.0  # WebSocket streaming for real-time data
websockets>=12.0,<14.0.0        # Async WebSocket for data provider
aiohttp>=3.9.0,<4.0.0           # Async HTTP for REST API calls

# =============================================================================
# UTILITIES
# =============================================================================
python-dotenv>=1.0.0,<2.0.0     # Environment variables
colorama>=0.4.6,<1.0.0          # Colored terminal output

# =============================================================================
# MONITORING (optional)
# =============================================================================
prometheus-client>=0.17.0       # Prometheus metrics export

# =============================================================================
# TESTING (install with: pip install -r requirements.txt[dev])
# =============================================================================
pytest>=7.0.0                   # Testing framework
